entire structure rework
