--- api/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "portfolio-api",
-  "version": "1.0.0",
-  "description": "Simple api for my portfolio website",
-  "main": "server.js",
-  "scripts": {
-    "start": "node src/server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Conor Manning",
-  "license": "ISC",
-  "dependencies": {
-    "body-parser": "^1.19.2",
-    "dotenv": "^16.0.0",
-    "ejs": "^3.1.6",
-    "express": "^4.17.3",
-    "serve-favicon": "^2.5.0"
-  }
-}
