--- api/src/views/_header.ejs
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-
-<head>
-    <script type="text/javascript" src="/static/index.js"></script>
-    <meta charset="utf-8">
-    <title>Conor Manning</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.7.2/font/bootstrap-icons.css">
-    <link rel="stylesheet" href="https://unpkg.com/bootstrap-table@1.21.1/dist/bootstrap-table.min.css">
-    <link rel="stylesheet" href="/static/glitch.css">
-</head>
