--- api/src/views/_nav.ejs
@@ -1,29 +1,0 @@
-<div class="container nav-container">
-    <nav class="navbar navbar-expand-sm navbar-light border-bottom">
-        <div class="container-fluid">
-            <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarsExample09"
-                aria-controls="navbarsExample09" aria-expanded="false" aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarsExample09">
-                <ul class="navbar-nav me-auto mb-2 mb-lg-0">
-                    <li class="nav-item">
-                        <a class="nav-link active" aria-current="page" href="/">Home</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="/about">About</a>
-                    </li>
-                    <li class="nav-item disabled">
-                        <a class="nav-link disabled">Work</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="/sandwiches">Sandwiches</a>
-                    </li>
-                </ul>
-            </div>
-                <a class="navbar-brand" href="/">
-                    <img src="/static/clubbin-logo.png" alt="Gone Clubbin" height="50">
-                </a>
-        </div>
-    </nav>
-</div>
