--- api/src/views/index.ejs
@@ -1,22 +1,0 @@
-<%- include('_header'); %>
-  <body>
-    <%- include('_nav'); %>
-    <div class="fakeHero container">
-      <div class="px-4 pt-5 mb-5 text-center">
-        <h1 class="center__text">Conor Manning</h1>
-        <h3>Software Dev. Data Analysis. Sandwich Recommendations.</h3>
-        <img class="d-block mx-auto pt-5" src="/static/clubbin-logo.png" alt="">
-      </div>
-      <div class="container">
-      </div>
-    </div>
- 
-<script>
-  document.body.getElementsByClassName('navbar-brand')[0].style.visibility = 'hidden' 
-</script>
-<script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.10.2/dist/umd/popper.min.js" integrity="sha384-7+zCNj/IqJ95wo16oMtfsKbZ9ccEh31eOz1HGyDuCQ6wgnyJNSYdrPa03rtR1zdB" crossorigin="anonymous"></script>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.min.js" integrity="sha384-QJHtvGhmr9XOIpI6YVutG+2QOK9T+ZnN4kzFN1RtK3zEFEIsxhlmWl5/YESvpZ13" crossorigin="anonymous"></script>
-
-  </body>
-  <%- include('_footer'); %>
-</html>
