+++ src/views/about.ejs
@@ -7,6 +7,7 @@
             <span id="age">0</span> year old Austin based developer who wants to automate everything. Currently working on integrations for B2B and DTC systems for mondoshop.com.
           </p>
           <p>
+<<<<<<< HEAD
             I strongly believe in proper data hygeine and accesibility when it comes to working with as many endpoints as we do these days.
             I am also a musician who is apparently obsessed with sandwiches and the religious documenting of my local food moves.
             Feel free to follow along through the sandwich products that I will be building out here on this site.
@@ -16,6 +17,20 @@
         </div>
     </div>
     <script src="/js/bootstrap.js"></script>
+=======
+            I strongly believe in proper data hygeine and accesibility when it comes to working with as many endpoints as we do these days,
+            I am also a musician who is apparently obsessed with sandwiches and the religious documenting of my local food moves.
+            Feel free to follow along through the sandwich products that I will be building out here on this site!
+          </p> 
+          <p>My cat has also regularly "helped" me with my projects, so if anything doesn't work it was probably his fault.</p>
+          <img class="wallace mx-auto d-block rounded" src="static/wallace_salad_cropped.jpg" alt="">
+        </div>
+    </div>
+    
+<script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.10.2/dist/umd/popper.min.js" integrity="sha384-7+zCNj/IqJ95wo16oMtfsKbZ9ccEh31eOz1HGyDuCQ6wgnyJNSYdrPa03rtR1zdB" crossorigin="anonymous"></script>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.min.js" integrity="sha384-QJHtvGhmr9XOIpI6YVutG+2QOK9T+ZnN4kzFN1RtK3zEFEIsxhlmWl5/YESvpZ13" crossorigin="anonymous"></script>
+
+>>>>>>> 5b4647f (entire structure rework)
   </body>
   <%- include('_footer'); %>
 </html>
